+++ Reservation/ReservationTableViewController.swift
@@ -7,9 +7,15 @@
 //
 
 import UIKit
+import CoreData
 
 class ReservationTableViewController: UITableViewController {
 
+    var mangedContext   = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
+    var branch: Branch?
+    var reservations    = [Reservation]()
+    var oldReservations = [Reservation]()
+    
     override func viewDidLoad() {
         super.viewDidLoad()
 
@@ -18,6 +24,31 @@ class ReservationTableViewController: UITableViewController {
 
         // Uncomment the following line to display an Edit button in the navigation bar for this view controller.
         // self.navigationItem.rightBarButtonItem = self.editButtonItem()
+    }
+    
+    override func viewWillAppear(_ animated: Bool) {
+        let fetchRequest   = NSFetchRequest<NSFetchRequestResult>(entityName: "Reservation")
+        let sortByFromDescriptor = NSSortDescriptor(key: "from", ascending: true)
+        let sortByToDescriptor   = NSSortDescriptor(key: "to",   ascending: true)
+        fetchRequest.sortDescriptors = [sortByFromDescriptor, sortByToDescriptor]
+        
+        // reservations
+        if let branch = self.branch {
+            let now = Date()
+            let predicate    = NSPredicate(format: "%K == %@ && %K > %@",  argumentArray: ["branch", branch, "to", now])
+            let oldPredicate = NSPredicate(format: "%K == %@ && %K <= %@", argumentArray: ["branch", branch, "to", now])
+            
+            do {
+                fetchRequest.predicate = predicate
+                reservations = try self.mangedContext.fetch(fetchRequest) as! [Reservation]
+                fetchRequest.predicate = oldPredicate
+                oldReservations = try self.mangedContext.fetch(fetchRequest) as! [Reservation]
+            } catch let error as NSError {
+                print("Cannot fetch CoreData: \(error), \(error.userInfo)")
+            }
+            
+            self.tableView.reloadData()
+        }
     }
 
     override func didReceiveMemoryWarning() {
@@ -28,68 +59,30 @@ class ReservationTableViewController: UITableViewController {
     // MARK: - Table view data source
 
     override func numberOfSections(in tableView: UITableView) -> Int {
-        // #warning Incomplete implementation, return the number of sections
-        return 0
+        return 2
     }
 
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        // #warning Incomplete implementation, return the number of rows
-        return 0
+        if section == 0 {
+            return reservations.count
+        }
+        return oldReservations.count
     }
 
-    /*
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "reuseIdentifier", for: indexPath)
-
-        // Configure the cell...
-
+        let cell = tableView.dequeueReusableCell(withIdentifier: "ReservationCell", for: indexPath)
+        let reservation: Reservation
+        
+        if indexPath.section == 0 {
+            reservation = reservations[indexPath.row]
+        } else {
+            reservation = oldReservations[indexPath.row]
+        }
+        
+        cell.textLabel?.text       = "\(reservation.from) - \(reservation.to)"
+        cell.detailTextLabel?.text = "\(reservation.reserver) (\(reservation.phone))"
+        
         return cell
     }
-    */
-
-    /*
-    // Override to support conditional editing of the table view.
-    override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-        // Return false if you do not want the specified item to be editable.
-        return true
-    }
-    */
-
-    /*
-    // Override to support editing the table view.
-    override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle, forRowAt indexPath: IndexPath) {
-        if editingStyle == .delete {
-            // Delete the row from the data source
-            tableView.deleteRows(at: [indexPath], with: .fade)
-        } else if editingStyle == .insert {
-            // Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
-        }    
-    }
-    */
-
-    /*
-    // Override to support rearranging the table view.
-    override func tableView(_ tableView: UITableView, moveRowAt fromIndexPath: IndexPath, to: IndexPath) {
-
-    }
-    */
-
-    /*
-    // Override to support conditional rearranging of the table view.
-    override func tableView(_ tableView: UITableView, canMoveRowAt indexPath: IndexPath) -> Bool {
-        // Return false if you do not want the item to be re-orderable.
-        return true
-    }
-    */
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destinationViewController.
-        // Pass the selected object to the new view controller.
-    }
-    */
 
 }