+++ CHANGELOG.md
@@ -3,4 +3,23 @@
 ---
 
 ###### 2017. 1. 2.
+
+BranchViewController:
+
+* selected branch is transferred to ReservationTableViewController when a row is selected.
+
+ReservationTableViewController:
+
+* reservations only related to selected branch was requested to CoreData.
+* following reservations and old reservations are divided.
+
+NewReservationViewController:
+
+* UIView generation for footerView happened to waste a lot of memory endlessly. Final footerView is never generated. The frame should only be replaced with keeping tableFooterView.
+* (Bug) required height is not correct. 
+
+---
+
+CocoaPod
+
 * CocoaPod has been installed.